Some builds queue

+++ src/Dzaba.TeamCitySimulator.Lib/Model/TimeEventData.cs
@@ -6,7 +6,7 @@ public sealed class TimeEventData
 {
     public DateTime Timestamp { get; set; }
     public string Name { get; set; }
-    public ushort QueueLength { get; set; }
+    public int QueueLength { get; set; }
     public uint TotalRunningBuilds { get; set; }
     public NamedQueueData[] AgentQueues { get; set; }
     public NamedQueueData[] BuildConfigurationQueues { get; set; }
@@ -15,5 +15,5 @@ public sealed class TimeEventData
 public sealed class NamedQueueData
 {
     public string Name { get; set; }
-    public ushort Length { get; set; }
+    public int Length { get; set; }
 }
